+++ src/main/java/de/maxhenkel/camera/Main.java
@@ -3,8 +3,8 @@ package de.maxhenkel.camera;
 import de.maxhenkel.camera.entities.ImageEntity;
 import de.maxhenkel.camera.entities.ImageRenderer;
 import de.maxhenkel.camera.gui.AlbumContainer;
-import de.maxhenkel.camera.gui.AlbumInventoryScreen;
 import de.maxhenkel.camera.gui.AlbumInventoryContainer;
+import de.maxhenkel.camera.gui.AlbumInventoryScreen;
 import de.maxhenkel.camera.gui.LecternAlbumScreen;
 import de.maxhenkel.camera.items.AlbumItem;
 import de.maxhenkel.camera.items.CameraItem;
@@ -13,16 +13,22 @@ import de.maxhenkel.camera.items.ImageItem;
 import de.maxhenkel.camera.net.*;
 import de.maxhenkel.corelib.ClientRegistry;
 import de.maxhenkel.corelib.CommonRegistry;
+import de.maxhenkel.corelib.tag.ItemTag;
+import de.maxhenkel.corelib.tag.SingleElementTag;
+import de.maxhenkel.corelib.tag.Tag;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.renderer.entity.EntityRenderers;
+import net.minecraft.core.HolderSet;
+import net.minecraft.core.Registry;
+import net.minecraft.core.RegistryAccess;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
-import net.minecraft.tags.ItemTags;
-import net.minecraft.tags.Tag;
+import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.item.crafting.SimpleRecipeSerializer;
 import net.minecraftforge.api.distmarker.Dist;
@@ -41,6 +47,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.lwjgl.glfw.GLFW;
 
+import java.util.Optional;
+
 @Mod(Main.MODID)
 public class Main {
 
@@ -58,7 +66,8 @@ public class Main {
     public static MenuType<AlbumInventoryContainer> ALBUM_INVENTORY_CONTAINER;
     public static MenuType<AlbumContainer> ALBUM_CONTAINER;
     public static EntityType<ImageEntity> IMAGE_ENTITY_TYPE;
-    public static Tag<Item> IMAGE_PAPER = ItemTags.bind(new ResourceLocation(Main.MODID, "image_paper").toString());
+    private static final ResourceLocation PAPER_LOCATION = new ResourceLocation(Main.MODID, "image_paper");
+    public static Tag<Item> IMAGE_PAPER;
 
     public static ServerConfig SERVER_CONFIG;
     public static ClientConfig CLIENT_CONFIG;
@@ -80,6 +89,15 @@ public class Main {
         CLIENT_CONFIG = CommonRegistry.registerConfig(ModConfig.Type.CLIENT, ClientConfig.class, true);
 
         DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> () -> FMLJavaModLoadingContext.get().getModEventBus().addListener(Main.this::clientSetup));
+        Registry<Item> blocks = RegistryAccess.BUILTIN.get().registryOrThrow(Registry.ITEM_REGISTRY);
+        TagKey<Item> itemTagKey = TagKey.create(Registry.ITEM_REGISTRY, PAPER_LOCATION);
+        Optional<HolderSet.Named<Item>> tag = blocks.getTag(itemTagKey);
+        if (tag.isPresent()) {
+            IMAGE_PAPER = new ItemTag(tag.get());
+        } else {
+            IMAGE_PAPER = new SingleElementTag<>(PAPER_LOCATION, Items.PAPER);
+            LOGGER.fatal("Failed to get image paper tag");
+        }
     }
 
     @SubscribeEvent