+++ changelog.md
@@ -1,2 +1,1 @@
-- Fixed empty ItemStack getting NBT data
-- Fixed can't take album out of empty album in lectern 
+- Added Norwegian translation