+++ changelog.md
@@ -1,1 +1,1 @@
-- Added Norwegian translation
+- Updated to 1.18.2